first commit -> see description
This is the first commit of the refactor. It includes many many changes that are all necessary to keep the app building and running.
Also, all functionality related to main process operations inside the renderer process are currently unavailable. This causes the app to be stuck at the loading screen upon launch. This is expected because IPC comms are moved to preload.js.